==> Notebooks/Notebook3-Enrichment_largeDeletions_BRCA1_2_inHRD.ipynb <==
*Ewing et al (2020) Structural variants at the BRCA1/2 loci are a common source of homologous repair deficiency in high grade serous ovarian carcinoma.*


# Notebook 3 - Are BRCA1/2 enriched for large deletions in HRD samples? 

This notebook processes and plots the results of tests of enrichment using circularised permutation to assess whether BRCA1/2 are enriched for large deletions in HRD samples. Figures included in Supplementary Figure 2. 

## Tabulate deletions in HRD samples

all_del_hrd_samps<-read.table("~/Desktop/BRCA1_BRCA2_SVs_paper/Analysis/all_hrd_samples_dels_sorted.bed",sep="\t")

library(reshape)
long.all_del_hrd_samps<-melt(all_del_hrd_samps,id.vars="V1")
long.all_del_hrd_samps<-long.all_del_hrd_samps[,c(1,3)]

chr_lengths<-read.table("hg38_chromosome_lengths.txt",sep="\t")
colnames(chr_lengths)[2]<-"value"
long.all_del_hrd_samps<-rbind(long.all_del_hrd_samps,chr_lengths)

long.all_del_hrd_samps<-long.all_del_hrd_samps[order(long.all_del_hrd_samps$V1,long.all_del_hrd_samps$value),]

## Calculate mean number of breakpoints from large deletions throughout the genome.

ordchrs<-c("chr1","chr2","chr3","chr4","chr5","chr6","chr7","chr8","chr9","chr10","chr11","chr12",
          "chr13","chr14","chr15","chr16","chr17","chr18","chr19","chr20","chr21","chr22","chrX")
long.all_del_hrd_samps$V1<-factor(long.all_del_hrd_samps$V1,levels=ordchrs)

chr_counts<-table(long.all_del_hrd_samps$V1)-1
chr_means<-(chr_counts/(chr_lengths[,2]))*5000000
chr_means

vline.data <- data.frame(V1 = c("chr13","chr13","chr17","chr17"),v2=c(32314862,32400266,43044295,43170245))
hline.data <- data.frame(V1 = levels(long.all_del_hrd_samps$V1),v2=as.numeric(chr_means))

## Make supplementary figure 2a

options(repr.plot.width=8,repr.plot.height=6)

ggplot(long.all_del_hrd_samps,aes(x=value))+
geom_histogram(binwidth=5000000,alpha=0.5)+
facet_wrap(~V1,scales="free_x")+
ylab("Large deletion breakpoint density")+xlab("Genomic position")+
 geom_vline(aes(xintercept = v2), vline.data,col="blue")+
geom_hline(aes(yintercept = v2), hline.data,col="black",lty="dashed")+
theme_bw()+
theme(axis.text.x = element_blank())
